# docs/openapi-v1.yaml
components:
  schemas:
    CalendarCreateRequest:
      properties:
        calendar:
          minLength: 1
          title: Calendar
          type: string
        end:
          title: End
          type: string
        location:
          anyOf:
          - type: string
          - type: 'null'
          title: Location
        notes:
          anyOf:
          - type: string
          - type: 'null'
          title: Notes
        start:
          title: Start
          type: string
        title:
          minLength: 1
          title: Title
          type: string
      required:
      - calendar
      - title
      - start
      - end
      title: CalendarCreateRequest
      type: object
    CalendarListResponse:
      properties:
        events:
          items:
            $ref: '#/components/schemas/_EventOut'
          title: Events
          type: array
      required:
      - events
      title: CalendarListResponse
      type: object
    CalendarUpdateRequest:
      properties:
        end:
          anyOf:
          - type: string
          - type: 'null'
          title: End
        location:
          anyOf:
          - type: string
          - type: 'null'
          title: Location
        notes:
          anyOf:
          - type: string
          - type: 'null'
          title: Notes
        start:
          anyOf:
          - type: string
          - type: 'null'
          title: Start
        title:
          anyOf:
          - type: string
          - type: 'null'
          title: Title
      title: CalendarUpdateRequest
      type: object
    ContactsSearchResponse:
      properties:
        contacts:
          items:
            $ref: '#/components/schemas/_ContactOut'
          title: Contacts
          type: array
      required:
      - contacts
      title: ContactsSearchResponse
      type: object
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
...
    IMessageInboundRequest:
      properties:
        agent:
          enum:
          - clu
          - tron
          - flynn
          title: Agent
          type: string
        body:
          minLength: 1
          title: Body
          type: string
        chat_guid:
          minLength: 1
          title: Chat Guid
          type: string
        from:
          minLength: 1
          title: From
          type: string
        received_at:
          minLength: 1
          title: Received At
          type: string
      required:
      - agent
      - from
      - body
      - received_at
      - chat_guid
      title: IMessageInboundRequest
      type: object
    IMessageInboundResponse:
      properties:
        event_id:
          title: Event Id
          type: string
        received:
          title: Received
          type: boolean
      required:
      - received
      - event_id
      title: IMessageInboundResponse
      type: object
    IMessageSendRequest:
      properties:
        body:
          minLength: 1
          title: Body
          type: string
        from:
          enum:
          - clu
          - tron
          - flynn
          - main
          title: From
          type: string
        service:
          default: iMessage
          enum:
          - iMessage
          - SMS
          title: Service
          type: string
        to:
          minLength: 1
          title: To
          type: string
      required:
      - from
      - to
      - body
      title: IMessageSendRequest
      type: object
    IMessageSentRequest:
      properties:
        agent:
          enum:
          - clu
          - tron
          - flynn
          title: Agent
          type: string
        body:
          minLength: 1
          title: Body
          type: string
        error_code:
          anyOf:
          - type: string
          - type: 'null'
          title: Error Code
        error_message:
          anyOf:
          - type: string
          - type: 'null'
          title: Error Message
        message_id:
          minLength: 1
          title: Message Id
          type: string
        sent_at:
          anyOf:
          - type: string
          - type: 'null'
          title: Sent At
        status:
          enum:
          - success
          - failed
          title: Status
          type: string
        to:
          minLength: 1
          title: To
          type: string
      required:
      - agent
      - message_id
      - to
      - body
      - status
      title: IMessageSentRequest
      type: object
    IMessageSentResponse:
      properties:
        acknowledged:
          title: Acknowledged
          type: boolean
        event_id:
          title: Event Id
          type: string
      required:
      - acknowledged
      - event_id
      title: IMessageSentResponse
      type: object
...
    ValidationError:
      properties:
        ctx:
          title: Context
          type: object
        input:
          title: Input
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
      type: object
...
    _ContactOut:
      properties:
        emails:
          items:
            type: string
          title: Emails
          type: array
        name:
          title: Name
          type: string
        phones:
          items:
            type: string
          title: Phones
          type: array
      required:
      - name
      - phones
      - emails
      title: _ContactOut
      type: object
...
    _EventOut:
      properties:
        calendar:
          title: Calendar
          type: string
        end:
          title: End
          type: string
        id:
          title: Id
          type: string
        location:
          anyOf:
          - type: string
          - type: 'null'
          title: Location
        notes:
          anyOf:
          - type: string
          - type: 'null'
          title: Notes
        start:
          title: Start
          type: string
        title:
          title: Title
          type: string
      required:
      - id
      - title
      - start
      - end
      - calendar
      title: _EventOut
      type: object
...
info:
  title: OpenClaw Bridge
  version: 1.0.0
openapi: 3.1.0
paths:
...
  /v1/calendar/events:
    get:
      operationId: list_events_v1_calendar_events_get
      parameters:
      - in: query
        name: from
        required: true
        schema:
          minLength: 1
          title: From
          type: string
      - in: query
        name: to
        required: true
        schema:
          minLength: 1
          title: To
          type: string
      - in: query
        name: calendar
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Calendar
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CalendarListResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: List Events
      tags:
      - calendar
    post:
      operationId: create_event_v1_calendar_events_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CalendarCreateRequest'
        required: true
      responses:
        '201':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Create Event
      tags:
      - calendar
  /v1/calendar/events/{event_id}:
    delete:
      operationId: delete_event_v1_calendar_events__event_id__delete
      parameters:
      - in: path
        name: event_id
        required: true
        schema:
          minLength: 1
          title: Event Id
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                title: Response Delete Event V1 Calendar Events  Event Id  Delete
                type: object
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Delete Event
      tags:
      - calendar
    patch:
      operationId: update_event_v1_calendar_events__event_id__patch
      parameters:
      - in: path
        name: event_id
        required: true
        schema:
          minLength: 1
          title: Event Id
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CalendarUpdateRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                title: Response Update Event V1 Calendar Events  Event Id  Patch
                type: object
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Update Event
      tags:
      - calendar
  /v1/contacts/search:
    get:
      operationId: search_contacts_v1_contacts_search_get
      parameters:
      - in: query
        name: q
        required: true
        schema:
          minLength: 1
          title: Q
          type: string
      - in: query
        name: limit
        required: false
        schema:
          default: 10
          maximum: 100
          minimum: 1
          title: Limit
          type: integer
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContactsSearchResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Search Contacts
      tags:
      - contacts
...
  /v1/imessage/inbound:
    post:
      operationId: imessage_inbound_v1_imessage_inbound_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IMessageInboundRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IMessageInboundResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Imessage Inbound
      tags:
      - imessage
  /v1/imessage/outbox:
    get:
      operationId: imessage_outbox_v1_imessage_outbox_get
      parameters:
      - in: query
        name: agent
        required: true
        schema:
          enum:
          - clu
          - tron
          - flynn
          title: Agent
          type: string
      - in: query
        name: timeout_s
        required: false
        schema:
          default: 25
          maximum: 60
          minimum: 0
          title: Timeout S
          type: integer
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Imessage Outbox
      tags:
      - imessage
  /v1/imessage/send:
    post:
      operationId: imessage_send_v1_imessage_send_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IMessageSendRequest'
        required: true
      responses:
        '202':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Imessage Send
      tags:
      - imessage
  /v1/imessage/sent:
    post:
      operationId: imessage_sent_v1_imessage_sent_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IMessageSentRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IMessageSentResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Imessage Sent
      tags:
      - imessage
...